backend/requirements.txt
fastapi==0.104.1
litellm==1.44.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
alembic==1.12.1
python-multipart==0.0.6